+++ platform/linux/window-dri.h
@@ -9,34 +9,31 @@
 #include <core/pixel.h>
 #include <core/shapes.h>
 #include <stdbool.h>
-#include <termios.h>
-#include <linux/fb.h>
+#include <xf86drmMode.h>
+
+struct drm_device {
+    i32 fd;
+    drmModeResPtr res;
+    drmModeConnectorPtr conn;
+    drmModeCrtcPtr crtc;
+
+    u32 width, height;
+    u32 refresh_rate;
+
+    bool initialized_;
+};
 
 struct window_dri {
-    i32 fd;
-    struct fb_fix_screeninfo fixed_info;
-    struct fb_var_screeninfo var_info;
+    struct drm_device dev;
 
-    void *mem;
-    u64 mem_size;
-
-    u32 xres, yres;
-    u32 padding;
-
-    rect_t win_rect;
-    rect_t display_rect;
-
-    bool closed;
-
-    i32 tty_fd;
-    struct termios orig_term_config;
+    struct gbm_device *gbm_dev;
 };
 
 i32 window_dri_open(struct window_dri *win,
     const rect_t *area, const u32 flags);
+
 void window_dri_close(struct window_dri *win);
 
-void window_dri_render_to_display(struct window_dri *win,
-    const struct pixel_flat_data *fb);
+void window_dri_render(struct window_dri *win, struct pixel_flat_data *fb);
 
 #endif /* P_WINDOW_DRI_H_ */